refactor(cli): extract sprint-status into its own module (#492)
sprint.py exceeded the 600-line module limit; cmd_sprint_status and
helpers moved to cli/sprint_status.py.
