+++ src/theforge/cli/main.py
@@ -16,6 +16,7 @@ from theforge.cli import (
     review,
     run,
     sprint,
+    sprint_status,
     status,
     telemetry,
 )
@@ -46,7 +47,7 @@ def build_parser() -> argparse.ArgumentParser:
     run.register_parser(subparsers)
     review.register_parser(subparsers)
     sprint.register_parser(subparsers)
-    sprint.register_sprint_status_parser(subparsers)
+    sprint_status.register_parser(subparsers)
     ideate.register_parser(subparsers)
     index.register_parser(subparsers)
     providers.register_parser(subparsers)
@@ -78,7 +79,7 @@ def main() -> None:
         "run": run.cmd_run,
         "review": review.cmd_review,
         "sprint": sprint.cmd_sprint,
-        "sprint-status": sprint.cmd_sprint_status,
+        "sprint-status": sprint_status.cmd_sprint_status,
         "ideate": ideate.cmd_ideate,
         "index": index.cmd_index,
         "check-providers": providers.cmd_check_providers,
